requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
sqlglot==23.0.2
scikit-learn==1.3.2
pytz==2023.3
pandas==2.1.4
numpy==1.26.2
